- Fix on Average() return type - Temp fix for Count(*) and SQLite

+++ sugar/DbLinq/Linq/Data/Sugar/Expressions/SpecialExpression.cs
@@ -63,7 +63,7 @@ namespace DbLinq.Linq.Data.Sugar.Expressions
             case SpecialExpressionType.Sum:
                 return defaultType; // for such methods, the type is related to the operands type
             case SpecialExpressionType.Average:
-                return typeof(decimal); // got no better idea
+                return typeof(double);
             case SpecialExpressionType.StringLength:
                 return typeof(int);
             case SpecialExpressionType.ToUpper:
